show pop up if somebrands have no product

+++ src/Components/Layout/Cart.jsx
@@ -1,22 +1,44 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { useLoaderData } from 'react-router-dom';
-
+import Swal from 'sweetalert2/dist/sweetalert2.js'
+import 'sweetalert2/src/sweetalert2.scss'
 const Cart = () => {
+
     const loaded = useLoaderData();
+    const [loadData, setLoadData] = useState(loaded)
     console.log(loaded)
     // const {_id, imageUrl, name, brandName, productType, price, rating, short_desc} = loaded;
     const handleDelete = _id => {
         console.log(_id)
-        fetch(`http://localhost:5000/cart/${_id}`, {
-            method: 'DELETE'
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/cart/${_id}`, {
+                    method: 'DELETE'
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data)
+                        Swal.fire(
+                            'Deleted!',
+                            'Your file has been deleted.',
+                            'success'
+                        )
+                        if (data.deletedCount > 0) {
+                            const remaining = loadData.filter(product => _id !== product._id)
+                            setLoadData(remaining);
+                        }
+                    })
+            }
         })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data)
-                // if (data.deletedCount > 0) {
-                //     const remaining = storedUsers.filter(user => _id !== user._id)
-                // }
-            })
+
     }
     let serial = 0;
     return (
@@ -35,7 +57,7 @@ const Cart = () => {
                         </tr>
                     </thead>
                     <tbody>
-                        {loaded?.map((cartItems, index) => (
+                        {loadData?.map((cartItems, index) => (
                             <tr key={cartItems._id}>
                                 <td className="hidden md:table-cell">{index + 1}</td>
                                 <td className="table-cell">{cartItems.name}</td>
